@@ -43,12 +43,12 @@
         "@types/jquery": "^3.5.4",
         "@types/leaflet": "^1.5.19",
         "@types/lodash": "^4.14.165",
-        "@types/node": "^14.14.6",
+        "@types/node": "^14.14.7",
         "@types/ua-parser-js": "^0.7.33",
         "@types/vue-select": "^2.5.1",
         "@types/webpack-env": "^1.15.3",
-        "@typescript-eslint/eslint-plugin": "^4.6.1",
-        "@typescript-eslint/parser": "^4.6.1",
+        "@typescript-eslint/eslint-plugin": "^4.7.0",
+        "@typescript-eslint/parser": "^4.7.0",
         "@vuepress/plugin-active-header-links": "^1.7.0",
         "@vuepress/plugin-back-to-top": "^1.7.0",
         "all-contributors-cli": "^6.19.0",
@@ -56,7 +56,7 @@
         "clean-webpack-plugin": "^3.0.0",
         "copy-dir": "^1.3.0",
         "css-loader": "^3.6.0",
-        "eslint": "^7.12.1",
+        "eslint": "^7.13.0",
         "eslint-plugin-prettier": "^3.1.4",
         "eslint-plugin-vue": "^7.1.0",
         "html-loader": "^1.3.2",
@@ -64,7 +64,7 @@
         "sass-loader": "^10.0.5",
         "string-replace-loader": "^2.3.0",
         "terser": "^5.3.8",
-        "ts-loader": "^8.0.9",
+        "ts-loader": "^8.0.11",
         "typescript": "^4.0.5",
         "url-loader": "^4.1.1",
         "vue-loader": "^15.9.5",
