@@ -250,13 +250,10 @@
                             :setting="setting"
                             v-model="settings[moduleId][settingId].value"
                             @input="update(moduleId, settingId)"
-                        >
-                            <template #titles>
-                                <component
-                                    :is="setting.titleComponent"
-                                ></component>
-                            </template>
-                        </settings-appendable-list>
+                            :module-id="moduleId"
+                            :setting-id="settingId"
+                            :orderable="!!setting.orderable"
+                        ></settings-appendable-list>
                         <pre v-else>{{ setting }}</pre>
                     </setting>
                 </div>
@@ -468,7 +465,11 @@ export default Vue.extend<
                         },
                     },
                     {
-                        title: this.$m('resetWarning.module'),
+                        title: this.$m('resetWarning.module', {
+                            module: this.$t(
+                                `modules.${this.modulesSorted[this.tab]}.name`
+                            ),
+                        }),
                         handler: () => {
                             Object.values(
                                 this.settings[this.modulesSorted[this.tab]]
