@@ -1,6 +1,9 @@
-export default (MODULE_ID: string, LSSM: Vue): unknown => ({
-    format: {
+import { ModuleSettingFunction } from 'typings/Module';
+import { Text } from 'typings/Setting';
+
+export default (() => ({
+    format: <Text>{
         type: 'text',
         default: 'DD.MM LTS',
     },
-});
+})) as ModuleSettingFunction;
