@@ -81,7 +81,7 @@ export default Vue.extend<
             ),
     },
     data() {
-        const buildingTypes = Object.values(this.$t('buildings')) as {
+        const buildingTypes = this.$t('buildings') as {
             [id: number]: InternalBuilding;
         };
         const categories = (this.$t('buildingCategories') as unknown) as {
