@@ -273,7 +273,9 @@ export default Vue.extend<
             this._setId(vehicle.id);
             this._setType('vehicles');
             this._setTitle(vehicle.caption);
-            this._setIcon('ambulance');
+            this._setIcon(
+                this.vehicleTypes[vehicle.vehicle_type].icon ?? 'ambulance'
+            );
             let additional = this.vehicleTypes[vehicle.vehicle_type].caption;
             if (vehicle.vehicle_type_caption)
                 additional += ` | ${vehicle.vehicle_type_caption}`;
