@@ -1,16 +1,19 @@
-export default (): unknown => ({
-    prevNextElement: {
+import { ModuleSettingFunction } from 'typings/Module';
+import { HotKey, Toggle } from 'typings/Setting';
+
+export default (() => ({
+    prevNextElement: <Toggle>{
         type: 'toggle',
         default: false,
     },
-    prevElementKey: {
+    prevElementKey: <HotKey>{
         type: 'hotkey',
         default: 'left',
         dependsOn: '.prevNextElement',
     },
-    nextElementKey: {
+    nextElementKey: <HotKey>{
         type: 'hotkey',
         default: 'right',
         dependsOn: '.prevNextElement',
     },
-});
+})) as ModuleSettingFunction;
