@@ -380,51 +380,51 @@ export default Vue.extend<
                 window.location.pathname.match(/\d+\/?/)?.[0] || '0'
             ),
             settings: {
-                title: true,
-                id: true,
-                type: true,
-                place: true,
-                prerequisites: true,
+                title: false,
+                id: false,
+                type: false,
+                place: false,
+                prerequisites: false,
                 vehicles: {
-                    title: true,
-                    content: true,
-                    patient_additionals: true,
+                    title: false,
+                    content: false,
+                    patient_additionals: false,
                     sort: 'caption',
                 },
                 chances: {
-                    normal: true,
-                    100: true,
+                    normal: false,
+                    100: false,
                 },
                 multifunctionals: {
-                    heavy_rescue_vehicles: true,
-                    battalion_chief_vehicles: true,
-                    platform_trucks: true,
-                    police_cars: true,
+                    heavy_rescue_vehicles: false,
+                    battalion_chief_vehicles: false,
+                    platform_trucks: false,
+                    police_cars: false,
                 },
                 optionalAlternatives: {
-                    allow_rw_instead_of_lf: true,
-                    allow_arff_instead_of_lf: true,
-                    allow_ktw_instead_of_rtw: true,
+                    allow_rw_instead_of_lf: false,
+                    allow_arff_instead_of_lf: false,
+                    allow_ktw_instead_of_rtw: false,
                 },
                 patients: {
-                    title: true,
-                    content: true,
-                    live: true,
-                    hideWhenNoNeed: true,
-                    patient_allow_first_responder_chance: true,
+                    title: false,
+                    content: false,
+                    live: false,
+                    hideWhenNoNeed: false,
+                    patient_allow_first_responder_chance: false,
                 },
                 prisoners: {
-                    title: true,
-                    content: true,
-                    live: true,
+                    title: false,
+                    content: false,
+                    live: false,
                 },
-                generatedBy: true,
-                credits: true,
-                expansions: true,
-                followup: true,
-                k9_only_if_needed: true,
-                hide_battalion_chief_vehicles: true,
-                bike_police_only_if_needed: true,
+                generatedBy: false,
+                credits: false,
+                expansions: false,
+                followup: false,
+                k9_only_if_needed: false,
+                hide_battalion_chief_vehicles: false,
+                bike_police_only_if_needed: false,
             },
             noVehicleRequirements: Object.values(
                 this.$m('noVehicleRequirements')
