@@ -3,6 +3,11 @@ import { InternalBuilding, ResolvedBuildingCategory } from '../Building';
 import { InternalVehicle, ResolvedVehicleCategory } from 'typings/Vehicle';
 import { Schooling } from 'typings/Schooling';
 
+interface ResolvedSchooling extends Pick<Schooling, 'caption' | 'duration'> {
+    required_for: string[];
+    [key: string]: string | string[];
+}
+
 export interface Overview {
     vehicles: {
         [id: number]: InternalVehicle;
@@ -42,7 +47,7 @@ export interface Overview {
         };
     };
     schoolingCategories: {
-        [category: string]: Schooling[];
+        [category: string]: ResolvedSchooling[];
     };
     schoolingsTab: {
         head: {
@@ -64,9 +69,9 @@ export interface OverviewComputed {
     currentBuildings: InternalBuilding[];
     buildingsFiltered: InternalBuilding[];
     buildingsSorted: InternalBuilding[];
-    currentSchoolings: Schooling[];
-    schoolingsFiltered: Schooling[];
-    schoolingsSorted: Schooling[];
+    currentSchoolings: ResolvedSchooling[];
+    schoolingsFiltered: ResolvedSchooling[];
+    schoolingsSorted: ResolvedSchooling[];
     vehicleTypes: InternalVehicle[];
 }
 
